intersection-observer/observer-exceptions.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
test(function () {
  assert_throws(RangeError(), function() {
    new IntersectionObserver(e => {}, {threshold: [1.1]})
  })
}, "IntersectionObserver constructor throws RangeError due to threshold=1.1.");

test(function () {
  assert_throws(RangeError(), function() {
    new IntersectionObserver(e => {}, {threshold: ["foo"]})
  })
}, "IntersectionObserver constructor throws RangeError due to threshold='foo'.");

test(function () {
  assert_throws("SYNTAX_ERR", function() {
    new IntersectionObserver(e => {}, {rootMargin: "1"})
  })
}, "IntersectionObserver constructor throws SYNTAX_ERR due to rootMargin='1'.");

test(function () {
  assert_throws("SYNTAX_ERR", function() {
    new IntersectionObserver(e => {}, {rootMargin: "2em"})
  })
}, "IntersectionObserver constructor throws SYNTAX_ERR due to rootMargin='2em'.");

test(function () {
  assert_throws("SYNTAX_ERR", function() {
    new IntersectionObserver(e => {}, {rootMargin: "auto"})
  })
}, "IntersectionObserver constructor throws SYNTAX_ERR due to rootMargin='auto'.");

test(function () {
  assert_throws("SYNTAX_ERR", function() {
    new IntersectionObserver(e => {}, {rootMargin: "1px 1px 1px 1px 1px"})
  })
}, "IntersectionObserver constructor throws SYNTAX_ERR due to invalid rootMargin='1px 1px 1px 1px 1px'.");

test(function () {
  assert_throws(TypeError(), function() {
    let observer = new IntersectionObserver(c => {}, {});
    observer.observe("foo");
  })
}, "IntersectionObserver.observe with an invalid target throws TypeError.");

</script>
